Slip timeseries_types in multiple files.
